requirements.txt
fitz==0.0.1.dev2
Pillow==9.5.0
Pillow==9.5.0
pynput==1.7.6
PySimpleGUI==4.60.5
websockets==11.0.3
